Extracted 'OnboardingOptionButton' View to it's own file
